Alternative queue-based source task
